refactor torch runner
