+++ python/orca/src/bigdl/orca/learn/pytorch/pytorch_ray_worker.py
@@ -31,130 +31,105 @@
 # This file is adapted from
 # https://github.com/ray-project/ray/blob/master/python/ray/util/sgd/torch/torch_runner.py
 
-
-from pyspark import BarrierTaskContext
-from pyspark.context import SparkContext
-from contextlib import closing
-import socket
+import ray
+from bigdl.orca.learn.pytorch.utils import find_free_port
 from bigdl.orca.learn.pytorch.torch_runner import TorchRunner
+import torch.nn as nn
 
 
-def find_free_port(tc):
-    address = tc.getTaskInfos()[tc.partitionId()].address.split(":")[0]
-    with closing(socket.socket(socket.AF_INET, socket.SOCK_STREAM)) as s:
-        s.bind(("", 0))
-        s.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
-        tc.barrier()
-        return f"{address}:{s.getsockname()[1]}"
+import logging
+logger = logging.getLogger(__name__)
+
+try:
+    from collections.abc import Iterable
+except ImportError:
+    from collections import Iterable
 
 
-def find_ip_and_port(pre_iter):
-    tc = BarrierTaskContext().get()
-    free_port = find_free_port(tc)
-    return [free_port]
-
-
-class TorchPysparkRunner(TorchRunner):
+class PytorchRayWorker(TorchRunner):
     """Manages a PyTorch model for training."""
 
     def __init__(self,
                  model_creator,
                  optimizer_creator,
-                 size,
-                 cluster_info,
-                 cores_per_worker,
                  loss_creator=None,
                  metrics=None,
                  scheduler_creator=None,
                  training_operator_cls=None,
                  config=None,
                  use_tqdm=False,
-                 scheduler_step_freq=None,
-                 state_dict=None,
-                 backend="torch-distributed",
-                 mode="fit"):
+                 scheduler_step_freq=None):
         super().__init__(model_creator, optimizer_creator, loss_creator, metrics, scheduler_creator,
                          training_operator_cls, config, use_tqdm, scheduler_step_freq)
 
-        self.state_dict = state_dict
-        self.size = size
-        self.mode = mode
-        self.backend = backend
-        self.cluster_info = cluster_info
+        self.backend = "torch-local"
+        self.rank = 0
+        self.size = 0
 
-        self.setup(cores_per_worker)
-        if self.backend == "torch-distributed":
-            self.setup_distributed(self.mode, self.cluster_info)
+    def setup_horovod(self):
+        import horovod.torch as hvd
+        hvd.init()
+        self.backend = "horovod"
+        self.rank = hvd.rank()
+        self.size = hvd.size()
+        self.setup_components_horovod()
+        self.setup_operator(self.models)
 
-    def setup_distributed(self, mode, cluster_info):
-        if mode == "fit":
-            self.rank = self._get_rank(cluster_info)
-            print("cluster is: ", cluster_info)
-            address = f"tcp://{cluster_info[0]}"
-            self.setup_torch_distribute(url=address,
-                                        world_rank=self.rank,
-                                        world_size=self.size)
+    def setup_address(self):
+        ip = ray._private.services.get_node_ip_address()
+        port = find_free_port()
+        return f"tcp://{ip}:{port}"
+
+    def setup_components_horovod(self):
+        import horovod.torch as hvd
+
+        logger.debug("Creating model")
+        self.models = self.model_creator(self.config)
+        if not isinstance(self.models, Iterable):
+            self.models = [self.models]
         else:
-            self.rank = 0
-            self.setup_components()
-            self.setup_operator(self.models)
+            raise ValueError("only support single model for now")
 
-    def _get_rank(self, cluster_info):
-        # As task placement may not be identical between two different jobs,
-        # we cannot simply index cluster_info using partitionId to get current
-        # ip and port.
-        # The approach here is to first get all tasks' ip in this job and compute
-        # a local rank by counting how many tasks has the same ip but with lower id.
-        # We then use the local rank to find the right slot in cluster_info to find
-        # the right global_rank.
-        tc = BarrierTaskContext().get()
-        infos = tc.getTaskInfos()
-        idx = tc.partitionId()
-        local_ip = infos[idx].address.split(":")[0]
-        local_rank = 0
-        for i in range(0, idx):
-            if infos[i].address.startswith(local_ip):
-                local_rank += 1
-        global_rank = -1
-        local_count = 0
-        for node in cluster_info:
-            if node.startswith(local_ip):
-                local_count += 1
-            global_rank += 1
-            if local_count == local_rank + 1:
-                break
-        return global_rank
+        assert all(isinstance(model, nn.Module) for model in self.models), (
+            "All models must be PyTorch models: {}.".format(self.models))
 
-    def train_epochs(self, data_creator, epochs=1, batch_size=32, profile=False,
-                     info=None, wrap_dataloader=None):
-        stats_list = super().train_epochs(data_creator, epochs, batch_size, profile, info,
-                                          wrap_dataloader)
-        state_dict = self.get_state_dict()
-
-        if self.rank == 0:
-            return [(state_dict, stats_list)]
+        logger.debug("Creating optimizer.")
+        self.optimizers = self.optimizer_creator(self.given_models,
+                                                 self.config)
+        if not isinstance(self.optimizers, Iterable):
+            hvd.broadcast_parameters(self.models[0].state_dict(), root_rank=0)
+            hvd.broadcast_optimizer_state(self.optimizers, root_rank=0)
+            parameters = self.models[0].named_parameters()
+            self.optimizers = hvd.DistributedOptimizer(self.optimizers,
+                                                       named_parameters=parameters)
+            self.optimizers = [self.optimizers]
         else:
-            return []
+            raise ValueError("only support one optimizer for now")
 
-    def validate(self, data_creator, batch_size=32, num_steps=None, profile=False,
-                 info=None, wrap_dataloader=None):
-        """Evaluates the model on the validation data set."""
-        validation_stats = super().validate(data_creator, batch_size, num_steps, profile, info,
-                                            wrap_dataloader)
-        if self.rank == 0:
-            return [validation_stats]
-        else:
-            return []
+        self._create_schedulers_if_available()
+        self._create_loss()
+
+    def get_node_ip(self):
+        """Returns the IP address of the current node."""
+        return ray._private.services.get_node_ip_address()
+
+    def find_free_port(self):
+        """Finds a free port on the current node."""
+        return utils.find_free_port()
 
     def predict(self, data_creator, batch_size=32, profile=False):
         """Evaluates the model on the validation data set."""
         config = self.config.copy()
         self._toggle_profiling(profile=profile)
 
-        partition = data_creator(config, batch_size)
+        shards_ref = data_creator(config, batch_size)
+        if not isinstance(shards_ref, ray.ObjectID):
+            raise ValueError("Only xshards is supported for predict")
+
+        partition = ray.get(shards_ref)
         return super().predict(partition=partition, batch_size=batch_size, profile=profile)
 
-    def get_state_dict(self):
+    def state_dict(self):
         """Returns the state of the runner."""
         state = {
             "epoch": self.epochs,
@@ -186,7 +161,7 @@ class TorchPysparkRunner(TorchRunner):
 
     def state_stream(self):
         """Returns a bytes object for the state dict."""
-        state_dict = self.get_state_dict()
+        state_dict = self.state_dict()
         _buffer = io.BytesIO()
         torch.save(state_dict, _buffer)
         return _buffer.getvalue()
@@ -205,7 +180,6 @@ class TorchPysparkRunner(TorchRunner):
 
     def shutdown(self):
         """Attempts to shut down the worker."""
-        dist.destroy_process_group()
         del self.training_operator
         del self.validation_loader
         del self.train_loader
